+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Application/Commands/Organization/GetOrganizationAuditLogsResponse.cs
@@ -12,13 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.Client.Models
-{
-    public sealed record ChangeOrganizationDto(
-        string Name,
-        string BusinessRegisterIdentifier,
-        AddressDto Address,
-        string? Comment,
-        OrganizationStatus Status,
-        string Domain);
-}
+using System.Collections.Generic;
+
+namespace Energinet.DataHub.MarketParticipant.Application.Commands.Organization;
+
+public sealed record GetOrganizationAuditLogsResponse(IEnumerable<OrganizationAuditLogDto> OrganizationAuditLogs);