+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Application/Commands/Organization/OrganizationAuditLogDto.cs
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.Client.Models
-{
-    public sealed record ChangeOrganizationDto(
-        string Name,
-        string BusinessRegisterIdentifier,
-        AddressDto Address,
-        string? Comment,
-        OrganizationStatus Status,
-        string Domain);
-}
+using System;
+using Energinet.DataHub.MarketParticipant.Domain.Model;
+
+namespace Energinet.DataHub.MarketParticipant.Application.Commands.Organization;
+
+public record OrganizationAuditLogDto(
+    Guid OrganizationId,
+    string Value,
+    Guid AuditIdentityId,
+    DateTimeOffset Timestamp,
+    OrganizationChangeType OrganizationChangeType);