+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Tests/Validation/UpdateOrganizationCommandRuleSetTests.cs
@@ -46,7 +46,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     string.Empty),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(Guid.Empty, organizationDto);
@@ -98,7 +99,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     "Denmark"),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
@@ -143,7 +145,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     "Denmark"),
                 "fake_value",
-                status);
+                status,
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
@@ -185,7 +188,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     "Denmark"),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
@@ -227,7 +231,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     "Denmark"),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
@@ -269,7 +274,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     value,
                     "Denmark"),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
@@ -311,7 +317,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     value),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
@@ -353,7 +360,8 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     "Denmark"),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
@@ -395,7 +403,51 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
                     string.Empty,
                     "Denmark"),
                 "fake_value",
-                "Active");
+                "Active",
+                "testDomain.dk");
+
+            var target = new UpdateOrganizationCommandRuleSet();
+            var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);
+
+            // Act
+            var result = await target.ValidateAsync(command);
+
+            // Assert
+            if (isValid)
+            {
+                Assert.True(result.IsValid);
+                Assert.DoesNotContain(propertyName, result.Errors.Select(x => x.PropertyName));
+            }
+            else
+            {
+                Assert.False(result.IsValid);
+                Assert.Contains(propertyName, result.Errors.Select(x => x.PropertyName));
+            }
+        }
+
+        [Theory]
+        [InlineData("", false)]
+        [InlineData(null, false)]
+        [InlineData("  ", false)]
+        [InlineData("testdomain.dk", true)]
+        public async Task Validate_Organizatiodomain_ValidatesProperty(string value, bool isValid)
+        {
+            // Arrange
+            var propertyName =
+                $"{nameof(UpdateOrganizationCommand.Organization)}.{nameof(ChangeOrganizationDto.Domain)}";
+
+            var organizationDto = new ChangeOrganizationDto(
+                ValidName,
+                ValidCvr,
+                new AddressDto(
+                    string.Empty,
+                    string.Empty,
+                    string.Empty,
+                    string.Empty,
+                    "Denmark"),
+                "fake_value",
+                "Active",
+                value);
 
             var target = new UpdateOrganizationCommandRuleSet();
             var command = new UpdateOrganizationCommand(_validOrganizationId, organizationDto);